Done: Auth, User-Student, & Learning Path

+++ app/Http/Controllers/LearningPathController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\LearningPath;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Str;
@@ -13,9 +14,7 @@ class LearningPathController extends Controller
     public function index()
     {
         $paths = LearningPath::all();
-        foreach($paths as $path){
-            $path->sum_courses = $path->courses()->count();
-        }
+        $paths->makeHidden(['created_at', 'updated_at']);
 
         return response()->json(['data' => $paths], 200);
     }
@@ -25,6 +24,11 @@ class LearningPathController extends Controller
      */
     public function store(Request $request)
     {
+        $user = Auth::user();
+        if(!$user->role === 'Superadmin'){
+            return response()->json(['error' => 'Unauthenticated.'], 401);
+        }
+
         $validator = Validator::make($request->all(), [
             'title' => 'required|string',
             'description' => 'required|string',
@@ -32,15 +36,15 @@ class LearningPathController extends Controller
         ]);
 
         if($validator->fails()){
-            return response()->json(['error' => $validator->errors()], 400);
+            return response()->json(['error' => $validator->errors()]);
         }
 
-        $field = $request->all();
+        $field = $request->only(['title', 'description']);
 
         if($request->hasFile('thumbnail_file')){
             $thumbnail = $request->file('thumbnail_file');
             $thumbnailPath = $thumbnail->storeAs(
-                'public/learning-paths',
+                '/public/learning-paths',
                 Str::slug($request->input('title')) . '_' . time() . '.' . $thumbnail->getClientOriginalExtension()
             );
 
@@ -49,16 +53,15 @@ class LearningPathController extends Controller
 
         LearningPath::create($field);
 
-        return response()->json(['msg' => 'Learning Path created successfully.'], 201);
+        return response()->json(['status' => true, 'message' => 'Berhasil menambahkan Learning Path.'], 201);
     }
 
     /**
      * Display the specified resource.
      */
-    public function show($id)
+    public function show($slug)
     {
-        $path = LearningPath::findOrFail($id);
-        $path->sum_courses = $path->courses()->count();
+        $path = LearningPath::where('slug', '=', $slug)->firstOrFail();
 
         return response()->json(['data' => $path], 200);
     }
@@ -66,9 +69,14 @@ class LearningPathController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, $id)
+    public function update(Request $request)
     {
-        $path = LearningPath::findOrFail($id);
+        $user = Auth::user();
+        if(!$user->role === 'Superadmin'){
+            return response()->json(['error' => 'Unauthenticated.'], 401);
+        }
+
+        $path = LearningPath::findOrFail($request->input('id'));
         $validator = Validator::make($request->all(), [
             'title' => 'required|string',
             'description' => 'required|string',
@@ -76,42 +84,47 @@ class LearningPathController extends Controller
         ]);
 
         if($validator->fails()){
-            return response()->json(['error' => $validator->errors()], 400);
+            return response()->json(['error' => $validator->errors()]);
         }
 
         $path->slug = null;
         $path->update($request->all());
 
         if($request->hasFile('thumbnail_file')){
-            if(Storage::exists($path->thumbnail) && $path->thumbnail !== 'public/learning-paths/thumbnail.png'){
+            if(Storage::exists($path->thumbnail) && $path->thumbnail !== '/public/learning-paths/thumbnail.png'){
                 Storage::delete($path->thumbnail);
             }
 
             $thumbnail = $request->file('thumbnail_file');
             $thumbnailPath = $thumbnail->storeAs(
-                'public/learning-paths',
+                '/public/learning-paths',
                 Str::slug($path->title) . '_' . time() . '.' . $thumbnail->getClientOriginalExtension()
             );
 
             $path->update(['thumbnail' => $thumbnailPath]);
         }
 
-        return response()->json(['data' => 'Learning Path updated successfully.']);
+        return response()->json(['status' => true, 'message' => 'Berhasil mengedit Learning Path.', 'data' => $path], 201);
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($id)
+    public function destroy(Request $request)
     {
-        $path = LearningPath::findOrFail($id);
+        $user = Auth::user();
+        if(!$user->role === 'Superadmin'){
+            return response()->json(['error' => 'Unauthenticated.'], 401);
+        }
 
-        if(Storage::exists($path->thumbnail) && $path->thumbnail !== 'public/learning-paths/default.png'){
+        $path = LearningPath::findOrFail($request->input('id'));
+
+        if(Storage::exists($path->thumbnail) && $path->thumbnail !== '/public/learning-paths/default.png'){
             Storage::delete($path->thumbnail);
         }
 
         $path->delete();
 
-        return response()->json(['msg' => 'Learning Path deleted successfully.'], 200);
+        return response()->json(['status' => true, 'message' => 'Berhasil menghapus Learning Path.', 'data' => $path], 200);
     }
 }
